Extended physics modelling capability by adding the high energy brem. model.
The high energy model (used above 1 GeV primary energy) for e-/e+ bremsstrahlung
photon emission has been added. This model is used, insead of the Seltzer-Berger one,
when the primary e-/e+ kinetic energy is above 1 GeV. The reason is that this model
includes the Landau–Pomeranchuk–Migdal suppression effect that can be important from
couple of GeV primary energies in acse of heavy (element) materials. The model is
based on an improved version of the Bethe-Heitler DCS with screening (TF model of the
atom) and Coulomb corrections. LPM and dielectric suppressions are also included.
The complete screening form of the DCS is used when the LPM effect is active or the
target atomic number is low (Z<5 i.e. TF model of atom is not appropriate).