Abstract RNG into an object with two function pointers
G4HepEmRandomEngine holds a reference to the real engine and two
function pointers to generate one random number or fill an array
with a given size, respectively. G4HepEmCLHEPRandomEngine is a
wrapper around CLHEP::HepRandomEngine and dispatches to the virtual
functions defined in the derived class.

The goal of this abstraction is to decouple the G4HepEmRun library
from CLHEP and pave the way for passing in custom RNG implementations
to the final state sampling. This enables further changes to reuse
more code on the GPU.