Factor out the direction computations for the brem. (SB, RB) interaction models.
The common part of secondary gamma and post interaction e-/e+ direction
computations of the SB and RB bremsstrahlung models are removed and put
into a separate function. This removed the corresponding duplicated code
and the makes the direction computations re-usable on the device.
The two models for performing the bremsstrahlung photon emission, i.e.
`PerformBremSB` and `PerformBremRB`, can now be (will be) merged into
a single function since all the differences are already removed into
the corresponding `SampleETransferBremX` methods.