Update CUDA build for CMake >= 3.17
To simplify building the CUDA parts of G4HepEM, bump minimum required
CMake version to 3.17 to take advantage of builtin support for CUDA
as a language and recent fixes. Older versions not yet supported due
to Geant4's multithreaded libraries exporting the -pthread flag which
some older versions of CMake may fail to wrap in `-Xcompiler` correctly
causing test builds to fail. Note that this can likely be resolved by
compiling C++/CUDA parts separately so that CUDA components do not
consume Geant4 usage requirements.

When CUDA enabled, check for and enable it as a language or fail. For
CMake < 3.18, use deprecated FindCUDA module to select appropriate
gencode flags. Add .cu sources directly to library builds, and for tests.