[tests] Remove kernels from header files
They are only called in the .cu files they are implemented in. Move
them to the top to make them declared before called. This fixes the
build with CUDA + tests enabled in an environment that doesn't have
device_launch_parameters.h in the include path, which now isn't used
anymore.