Only use CUDA Runtime in libg4HepEMData
All CUDA calls in g4HepEM are to the CUDA runtime API, and so NVCC is
not required to compile the code nor to have separate .cu files.

Move all host-device copy/free functions from .cu files to the
corresponding .cc files, protecting compilation with the needed ifdef.
Find the CUDA toolkit libraries when CUDA enabled, linking the runtime
API library to g4HepEmData as a private dependency.