Rationalize test and example applications
Move all pure tests under testing subdirectory for clarity and
simplicity as these do not need to be full CMake projects.

Build TestUtils as a common static library to simplify build and use,
linking tests to this. Build each test program in isolation using
standard CMake commands for building, linking, and adding as a test.