Adding gamma interactions to G4HepEm and activate them.
This is a huge commit with many a large amount of new functionality
for making G4HepEm cover gamma interactions. Interactions covered
are:
  - production of e- e+ pairs (final form)
  - Compton scattering (final form)
  - photoelectric absorption (simplified)

All gamma interactions are in their final form except the photoelectric
absorption. This will require a bit more time later but a simple
absorption of gammas (when teir energy drops below 250 [keV]) has
been implemented in order to prevent the unrealisticaly high, low
energy e- production due to the high number of Compton scattering while
the photon leaves the target or its enegry drops below 100 [eV] in a
Compton scatteirng. Note, that all these unrealistic behaviour is the
consequence of the missing of photoelectric absorption, that would absorb
the gamma when its energy is getting lower with a higher probability.

The corresponding gamma manager, with its `Howfar` and `Perfom` functions
has been also implemented and active now in G4HepEm.