Update the `TestEm3` example applications according the new, gamma simulation capabilities.
The physics lists have been updated and includes now the gamma
interactions. A `GammaAbsorption` process, as a `G4VDisreteProcess`
has been implemented and used in the Geant4 EM physics lists instead
of the photoelectric process. This is for having the same gamma
absorption, when using the native Geant4 EM physics, as in G4HepEm.
This will be updated when G4HepEm has the final version of the cor-
responding photoelectric absorption implenented.