G4HepEmState struct to associate parameters and data
Many G4HepEmRun functions take both G4HepEmData and G4HepEmParameters
for their calculations. Clients serializing data will also need to store
the parameters with the data for consistency when using the deserialized
data.

Add new G4HepEmState struct that is a simple association between pointers to a
G4HepEmData instance and a corresponding G4HepEmParameters instance. This
is implemented as a simple view of these pointers, and no (de)allocation
is performed or implemented.

Add JSON (de)serialization functions for G4HepEmParameters and G4HepEmState.

Update DataImportExport test to roundtrip G4HepEmState and the held parameters
and data.