Fix data transfer to GPU and make it more resistent
When implementing the (simplified) photoelectric effect in commit
26f6ead338, I assumed that non-pointer members are automatically
copied to the GPU via a memcpy. As this was not the case, the new
data members were missing and not properly initialized. Change all
copy functions to now use a memcpy of the data structure to catch
newly added fields in the future.