Avoid globbing in CMake files
This is gennerally not recommended because adding or deleting a file
doesn't retrigger CMake configuration, so the list of sources it not
updated and the build fails (either due to missing files or missing
symbols in files that were not compiled). Furthermore explicit lists
are needed to conditionally add files for tracking managers when
compiling against Geant 11.0 and later.