Move energy loss fluctuations after MSC
This simplifies the interplay of the two because MSC works on the
mean energy loss. The minor disadvantage is that the check for the
tracking cut happens after MSC, which means that a particle might
scatter only to be killed afterwards. This is very rare and not of
pratical relevance in terms of performance, but leads to minor
changes even in case loss fluctuations are turned off.