Add possibility of MSC stepping to HepEmTrackingManager
If a step is only limited by MSC, the HepEmTrackingManager can re-try
the step with the remaining distance to the discrete interaction,
without applying energy loss fluctuations and re-computing the cross
sections. This option is enabled by default because it provides a small
performance benefit, but can be turned off to get results identical
to the G4HepEmProcess.