Restore comparison of Urban MSC before Geant4 11.0
Geant4 11.0 contained a large refactoring of G4PhysicsVector, which
slightly altered the numerical value of the energy bins. This should
not be a problem, but the transport mean free path computed by the
Urban model is not continuous around 10 MeV, so we have to pick the
"right" side for each version to obtain as close results as possible.