Allow CMAKE_CXX_STANDARD to be set by user
Experiments may wish to compile all dependencies with the same C++
standard used in all their other code for hard ABI compatibility. We
may also wish to pick up an existing standard if G4HepEm is ever used
as a subproject - as in Adept.

Only set CMAKE_CXX_STANDARD to minimum required if it is not already
set. Fail if this does not meet the required min of 11.