Create function for creation of logarithmic energy grids
During G4HepEm's initialization, several data tables for gammas and
electrons/positrons are calculated using a grid of energies with
points between endpoints uniformly spaced in log(E). In general, the
calculation only requires the endpoints and the total number of
points, G4HepEm also storing log(Emin) and the inverse of the
spacing in log(E) for later convenience. The code required for
the filling of these energy grids is common, but duplicated for each
place they are needed.

Implement the code to fill a logarithmicly-spaced array of points
in a helper function in G4HepEmInitUtils, taking a pre-existing array
as input. Return the filled array together with the log of the lower
bound, and inverse of the spacing in log(E).

Refactor consumers of log-energy grids in G4HepEmInit to use the new
function.

Fixup recommited code