Support building G4HepEm without Geant4
Use cases for G4HepEm not requiring use of Geant4 have recently been
identified, only needing the data and run components. Generation of
data would still be required, but can be persisted separately and
read in without needing Geant4 via G4HepEm's JSON persistency.

Add CMake option to enable support for Geant4, defaulting to ON to
match current behaviour. Update build so that when this is OFF, it

- Does not build the G4HepEm and G4HepEmInit components
- Finds the CLHEP package to support random number generation in
  G4HepEmRun
- Sets the G4VERSION_NUM to 1100 (TODO: make selectable)
- Add a "geant4" component in installed CMake config file to allow
  detection of this feature
- Adapt G4HepEm_LIBRARIES for both build variants
- Do not build tests even when testing enabled as they currently
  all require Geant4 or G4HepEm's support for it