Include G4 fast simulation processes into the tracking manager.
Fast simulation manager processes, attached to the e-/e+ and gamma
by the user, were ignored so far in the tracking manager implementation
of the stepping loop. These are included now. The gamma tracking has
been reimplemented without using the helper in order to be able to call
the fast simulation process and handle that properly during the stepping.