Including electron/positron nuclear interaction handling to the HepEm tracking manager.
Similarly to the gamma-nuclear, the native Geant4 electron/positron nuclear processes
are used to perform the interaction when happens. The Geant4 processes are looked for
in the electron/positron process manager at the init of the tracking manager. A pointer
to the Geant4 process is stored in the tracking manager that allows to invoke the interaction
if needed. If there are no nuclear interaction attached to the electron/positron, then the
local pointer to the nuclear process stays nullptr and no interaction is invoked in that case.
So everything works fine even without nuclear processes attached to electron/positron by
the user in the Geant4 physics list.