Move to a new representation of the macroscopic cross section data in case of gamma.
Macroscopic cross section related data for gamma are stored now in a new format
that supports the usage of a single, total macroscopic cross section (the sum of
the macroscopic cross section of the possible, individual interactions). This can
be obtained with a single interpolation and can be used to sample now the step length
till the next interaction where the type of the interaction can be selected based on
the individual process cross sections that are also available.

Note: relying on a single cross section evaluation at the step limit phase can save up
so cross section evaluations/interpolations when the step was limited by volume boundary
and interaction not happens. This is because then we do not need the infomation on the
process type.