Introducing the possiblity of configuration per detector region.
Allows to use different simulation configuration in the different
detector regions. This is achived by storing and allowing to set
some of simulation parameters per detector regions. The values
of the parameters are red at rung time depending on the actual
region.

The following parameters can be set per detector region (at the moment):
 - MSC step limit parameters: range and safety factors or default/minimal
   step limit algorithm (the minimal step limit the algorithm is still not in)
   (this is required for CMS and LHCb)

 - parameters of the continuous energy loss step limit function (these are
   the dRoverRange and finalRange)
   (this is required for LHCb)

 - flag to indicate if the energy loss fluctuation is active (ON/OFF)
   (this will be required by ATLAS)

 - activating Woodcock-tracking of photons (ON/OFF)
   (this will be used by ATLAS, CMS and others)

 - flag to indicate if the combined MSC with Transportation is allowed through
   multiple steps
   (this will be used by ATLAS)