An additional paramater to control if the e+ correction to MSC angle should be applied.
A global flag has been added to control if the e+ correction to the theta_0 MSC
angle should be applied or not. It's ON (true) by default, i.e. there is e+
correction, but should be set to OFF (false) in the ATLAS simulation (as they
don't use it: responsible for moving away from their experimental data).