Remove the G4VProcess implementation of the G4HepEm physics.
Using the G4VProcess implementaion of the G4HepEm physics, as the connection
to any Geant4 application, was a convenient way of testing the correctness of
the physics during the developemnt.

The complete and most performant connection to Geant4 today comes with the
G4HepEmTrackingManager implementation. This includes the G4HepEm physics but
also the tracking/stepping specialised for e-/e+, gamma as well as the special
tracking algorithms, e.g. Woodock tracking for photons and the combined MSC and
transportation for e-/e+. Moreover, gamma- and leton-nuclear interaction as
well as fast simulation processes are also incorporated into this tracking
manager (in case these have been attached to e-/e+ or gamma in the Geant4
physics list utilised in the Geant4 application.

Therefore, enforcing the usage of the G4HepEmTrackingManager ensures to
provide complete functionalites combined with maximal perfrormance benefits.

(The process implementation kept in the example application for backward compatibity.)