Include also lepton-nuclear interactions into a single method.
The `PerformNuclear` method can be used to invoke the gamma/electron/positron - nuclear
interactions now depending on the `particleID` input parameter value (2/0/1 for gamma/
electron/positron - nuclear interaction). The method is self contained, i.e. everything
that is needed to invoke the corresponding `G4HadronInelasticProcess` is done inside
the method. The interaction is invoked, the step and track are updated to their post-
interaction state, the secondaries are stacked to the secondary track vector of the
input step and the energy deposited during the interaction (or due to applying the cut)
is returned.