Make sure that the example application CMakeLists works as intended.
We check if the tracking manager is supported (by the utilised Geant4
version) by using a small cxx code when configuring and installing G4HepEm.
The G4HepEm_HAS_G4VTRACKINGMANAGER variable is set to True if the code
compiles while the G4HepEm_trackingmanager_FOUND variable goes to the
G4HepEm config file with a True value. The first is availabel in the same
unit, i.e. when building the example application together with G4HepEm
e.g. during testing, while only the other is available when using an
already installed G4HepEm version. This modification ensures that the
application builds as expected in both cases.