Adding two EM parameters.
An EM parameter, that corresponds to the G4 LateralDisplacement, has been
added to the HepEm parameters. It's initial value is taken from the
G4EmParameters at init and controls if lateral displacement is computed
in MSC. It allows a bit of optimisation in case of large volumes and
used by default in the option2 G4 EM physics constructor (by LHCb).

The additional parameter is a tracking cut for gamma which do not exist
in G4 in this form. It behaves similarly to the corresponding tracking cut
for electron: the gamma tracking is stoped when its energy drops below this
energy limit and the current kinetic energy is deposited at that point.
This can be achived in several ways inside G4 (e.g. user stepping action or
G4UserLimits, etc.). Such a tracking cut is applied by LHCb and this is
why it has been introduced now. The default value is zero, i.e. no any
tracking cut for gamma by default, that can be changed when constructing
the tracking manager.